+++ js/main.js
@@ -1,1 +1,4 @@
-import './create-array-photo.js';
+import {createPhotos } from './create-array-photo.js';
+
+const photos = createPhotos();
+console.log({photos});
